+++ dspace-api/src/main/java/org/dspace/content/clarin/ClarinLicenseServiceImpl.java
@@ -1,25 +1,20 @@
 package org.dspace.content.clarin;
 
+import java.sql.SQLException;
+import java.util.List;
+import java.util.Objects;
+
 import org.apache.commons.lang.NullArgumentException;
-import org.checkerframework.checker.units.qual.C;
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.authorize.service.AuthorizeService;
 import org.dspace.content.dao.clarin.ClarinLicenseDAO;
 import org.dspace.content.service.clarin.ClarinLicenseService;
-import org.dspace.core.Constants;
 import org.dspace.core.Context;
 import org.dspace.core.LogHelper;
-import org.dspace.eperson.Group;
-import org.dspace.eperson.GroupServiceImpl;
-import org.dspace.event.Event;
 import org.hibernate.ObjectNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-
-import java.sql.SQLException;
-import java.util.List;
-import java.util.Objects;
 
 public class ClarinLicenseServiceImpl implements ClarinLicenseService {
 
@@ -63,26 +58,41 @@ public class ClarinLicenseServiceImpl implements ClarinLicenseService {
     }
 
     @Override
-    public List<ClarinLicense> findAll(Context context) throws SQLException {
+    public List<ClarinLicense> findAll(Context context) throws SQLException, AuthorizeException {
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                     "You must be an admin to create an Clarin License");
+        }
+
         return clarinLicenseDAO.findAll(context, ClarinLicense.class);
     }
 
 
     @Override
-    public void delete(Context context, ClarinLicense clarinLicense) throws SQLException {
+    public void delete(Context context, ClarinLicense clarinLicense) throws SQLException, AuthorizeException {
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "You must be an admin to create an Clarin License");
+        }
+
         clarinLicenseDAO.delete(context, clarinLicense);
     }
 
     @Override
-    public void update(Context context, ClarinLicense newClarinLicense) throws SQLException {
+    public void update(Context context, ClarinLicense newClarinLicense) throws SQLException, AuthorizeException {
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "You must be an admin to create an Clarin License");
+        }
+
         if (Objects.isNull(newClarinLicense)) {
             throw new NullArgumentException("Cannot update clarin license because the new clarin license is null");
         }
 
-        ClarinLicense foundClarinLicense = find(context, newClarinLicense.getId());
+        ClarinLicense foundClarinLicense = find(context, newClarinLicense.getID());
         if (Objects.isNull(foundClarinLicense)) {
-            throw new ObjectNotFoundException(newClarinLicense.getId(), "Cannot update the license because the clarin license wasn't found " +
-                    "in the database.");
+            throw new ObjectNotFoundException(newClarinLicense.getID(),
+                    "Cannot update the license because the clarin license wasn't found in the database.");
         }
 
         clarinLicenseDAO.save(context, newClarinLicense);
